start audio test

+++ src/Tests/TestBubbleSortVis.java
@@ -44,9 +44,9 @@ public class TestBubbleSortVis extends Test {
         super.OnRender();
         for (int i = 0; i < arr.length; i++) {
             if(sorted)
-                renderer.fillRect(new Vector2f( -Window.dim.x/2f + i * Window.dim.x/2f/ arr.length*2, -Window.dim.y/2f), new Vector2f(Window.dim.x/2f/ arr.length, arr[i] * Window.dim.y/ arr.length), new Vector4f(0, 1, 0,1));
+                renderer.fillRect(new Vector2f( -Window.dim.x/2f + i * Window.dim.x/2f/ arr.length*2, -Window.dim.y/2f), new Vector2f(Window.dim.x/2f/ arr.length, (float) (arr[i] * Window.dim.y) / arr.length), new Vector4f(0, 1, 0,1));
             else
-                renderer.fillRect(new Vector2f( -Window.dim.x/2f + i * Window.dim.x/2f/ arr.length*2, -Window.dim.y/2f), new Vector2f(Window.dim.x/2f/ arr.length, arr[i] * Window.dim.y/ arr.length));
+                renderer.fillRect(new Vector2f( -Window.dim.x/2f + i * Window.dim.x/2f/ arr.length*2, -Window.dim.y/2f), new Vector2f(Window.dim.x/2f/ arr.length, (float) (arr[i] * Window.dim.y) / arr.length));
         }
     }
 }
